+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Unreleased
+    - CART can now print text representation of decision rules
     - LOF and Isolation Forest contamination off by default
     - Embed method now returns an array instead of dataset
     - Fixed issue with Dataset partitioning